Add Full Size Image Event And Conditional
This commit sets up the nfrastructure for a future function that will show the full sized image in a new window.  Currently it is not possible to use the selected image as it has to be saaved to the server first.  Hopefully once this code is integrated, it will be possible to use this code effectively once it is integrated into CommCare.